refactor: session null check 부분 수정
