+++ tomcat/src/main/java/org/apache/catalina/SessionManager.java
@@ -2,6 +2,7 @@ package org.apache.catalina;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import org.apache.coyote.http11.common.Session;
 
 public class SessionManager implements Manager {
 
@@ -9,12 +10,15 @@ public class SessionManager implements Manager {
 
     @Override
     public void add(final Session session) {
+        if (session == null) {
+            return;
+        }
         sessions.put(session.getId(), session);
     }
 
     @Override
     public Session findSession(final String id) {
-        return sessions.get(id);
+        return sessions.getOrDefault(id, new Session());
     }
 
     @Override
