+++ tomcat/src/main/java/org/apache/coyote/http11/common/Session.java
@@ -1,13 +1,18 @@
-package org.apache.catalina;
+package org.apache.coyote.http11.common;
 
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.UUID;
 
 public class Session {
 
     private final String id;
     private final Map<String, Object> items = new HashMap<>();
+
+    public Session() {
+        this(UUID.randomUUID().toString(), Collections.emptyMap());
+    }
 
     public Session(final String id) {
         this(id, Collections.emptyMap());
